Add additional_repositories option to source configuration
Fix the use case where there are multiple local private repositories.
example: one repository is hosting an npm mirror, and the other has
private packages which depend on packages from npm.

Currently this case will fail when publishing because npm attempts to
validate the package dependencies and fails when it can't find the
public packages in the private repository.